Bug fix for handling split and combined FDs
Previously the logic for determining if an FD exists in a collection of
FDs was to simply compare the string output of the FDs. But this doesn't
work if the collection has the split FDs (i.e. A-> B and A-> C) and the
input FD is the combined form (A-> B, C). This has been corrected.